update plotting routines to use metadata if available
